requirements.txt
absl-py==2.0.0
appdirs==1.4.4
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cryptography==41.0.3
cycler==0.12.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
filelock==3.9.0
fonttools==4.43.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.18.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
MarkupSafe==2.1.1
matplotlib
mpmath==1.3.0
munch==4.0.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pathtools==0.1.2
Pillow==10.0.1
pip==23.3
piqa==1.3.2
pretrainedmodels==0.7.4
protobuf==4.23.4
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
segmentation-models-pytorch==0.3.3
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.0.0
six==1.16.0
smmap==5.0.1
sympy==1.11.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
torch==2.1.0
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.7.1
urllib3==1.26.16
wandb==0.15.12
Werkzeug==3.0.1
wheel==0.41.2